% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_appendix_section.R
\name{document_appendix_section}
\alias{document_appendix_section}
\title{Generate an "Appendix" Section for Groups}
\usage{
document_appendix_section(
  data,
  grouped_by = "Species",
  country_code_col = "CountryCode",
  region_col = "Region",
  locality_col = "Locality",
  lat_col = "Lat",
  long_col = "Long",
  elevation_col = "Elevation",
  institution_col = "Institution",
  number_col = "Number"
)
}
\arguments{
\item{data}{A data frame containing specimen data with columns for grouping, country, region, locality,
latitude, longitude, elevation, institution, and specimen number.}

\item{grouped_by}{Column name to group specimens by (default: \code{"Species"}).}

\item{country_code_col}{Column name for country code (default: \code{"CountryCode"}).}

\item{region_col}{Column name for region information (default: \code{"Region"}).}

\item{locality_col}{Column name for locality information (default: \code{"Locality"}).}

\item{lat_col}{Column name for latitude values (default: \code{"Lat"}).}

\item{long_col}{Column name for longitude values (default: \code{"Long"}).}

\item{elevation_col}{Column name for elevation values (default: \code{"Elevation"}), assumed to be in meters above sea level.}

\item{institution_col}{Column name for institution information (default: \code{"Institution"}).}

\item{number_col}{Column name for specimen number (default: \code{"Number"}).}
}
\value{
A formatted appendix section with specimen details grouped by the specified column. Missing values or columns will be printed as \code{"N/A"}.
}
\description{
This function generates a formatted appendix section with specimen details grouped by a specified column.
Missing columns in the dataset will be automatically filled with \code{"N/A"}, ensuring the function runs without errors and retains proper formatting, however, the function performs optimally when all specified columns are present in the dataset.
}
\examples{
data(data_morphodata)
document_appendix_section(
  data = data_morphodata,
  grouped_by = "Species",
  country_code_col = "Country",
  lat_col = "Latitude",
  long_col = "Longitude")

}
