% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_description_section.R
\name{document_description_section}
\alias{document_description_section}
\title{Generate "Description" Sections for Groups}
\usage{
document_description_section(data, trait_columns, grouped_by = "Species")
}
\arguments{
\item{data}{A data frame containing specimen data, with columns for group designations (e.g., "Species", "Clade", ect.) and trait values.}

\item{trait_columns}{A character vector of column names corresponding to the traits to summarize.}

\item{grouped_by}{A character string specifying the column name used to group specimens (default: \code{"Species"}).}
}
\value{
A character vector of group descriptions, each formatted to list the traits and their corresponding range or value.
}
\description{
This function generates description sections for each group based on the minimum and maximum values for a set of traits. The descriptions are formatted for each group, listing trait ranges or single values.
}
\details{
The function computes the minimum and maximum values for each trait and constructs a description for each group that includes the ranges or single values of the traits. If a trait has missing values for all individuals in a group, it will be marked as \code{"N/A"}. Ranges are reported as \code{"min-max"}, and single values are reported as is. Categorical and binary traits will be printed as a range if multiple values are present within the same species.
Prior to publishing, all en dashes in ranges should be replaced with em dashes.
}
\examples{
data(data_morphodata)
document_description_section(
  data = data_morphodata,
  trait_columns = c("trait1", "trait2", "trait3", "trait4"),
  grouped_by = "Clade")

}
