% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_diagnoses_section.R
\name{document_diagnoses_section}
\alias{document_diagnoses_section}
\title{Generate "Diagnoses" Sections Based on Trait Ranges}
\usage{
document_diagnoses_section(data, trait_columns, grouped_by)
}
\arguments{
\item{data}{A data frame containing specimen data, with columns for group designations (e.g., "Species", "Clade", ect.) and trait values.}

\item{trait_columns}{A character vector of column names corresponding to the traits to be analyzed for non-overlapping ranges.}

\item{grouped_by}{A character string specifying the column name used to group specimens.}
}
\value{
A list of character strings, each representing a description that distinguishes two groups based on non-overlapping trait ranges. If no non-overlapping ranges are found, a message is printed.
}
\description{
This function generates diagnostic descriptions distinguishing groups based on non-overlapping ranges of specified traits. It compares the minimum and maximum values for each trait across groups and produces a summary of traits where the ranges do not overlap.
}
\details{
The function calculates the minimum and maximum values for each trait by group, compares these ranges across groups, and generates descriptions for trait comparisons that do not overlap. If no non-overlapping ranges are found, a message will indicate this and list the groups that could not be separated.
Categorical and binary traits will be printed as a range if multiple values are present within the same group. Prior to publishing, all en dashes in ranges should be replaced with em dashes.
}
\examples{
data(data_morphodata)
document_diagnoses_section(
  data = data_morphodata,
  trait_columns = c("trait1", "trait2"),
  grouped_by = "Species")

}
