% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_get_taxonomy.R
\name{document_get_taxonomy}
\alias{document_get_taxonomy}
\title{Fetch Taxonomic Hierarchy from NCBI}
\usage{
document_get_taxonomy(name)
}
\arguments{
\item{name}{A character string specifying the name to query in the NCBI taxonomy database.}
}
\value{
A named vector where:
\itemize{
\item \code{Names} The taxonomic ranks.
\item \code{Values} The corresponding scientific names for each rank.
}
}
\description{
This function retrieves the taxonomic hierarchy for a given taxonomic name from the NCBI taxonomy database.
It returns a named vector where each element represents a taxonomic rank and its corresponding scientific name.
}
\details{
The function queries the NCBI taxonomy database using \code{entrez_search} to retrieve group
information. It then fetches taxonomic details using \code{entrez_fetch}, parses the XML response, and
extracts the hierarchical taxonomic ranks and their corresponding names.
}
\examples{
document_get_taxonomy(name = "Homo sapiens")

}
