% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_material_examined.R
\name{document_material_examined}
\alias{document_material_examined}
\title{Generate a "Material Examined" Section for Taxonomic Descriptions}
\usage{
document_material_examined(
  data,
  grouped_by = "Species",
  region_col = "Region",
  institution_col = "Institution",
  number_col = "Number",
  collector_col = "Collector",
  locality_col = "Locality",
  date_collected_col = "DateCollected",
  type_col = "Type",
  sex_col = "Sex",
  lat_col = "Lat",
  long_col = "Long",
  elevation_col = "Elevation"
)
}
\arguments{
\item{data}{A data frame containing specimen metadata and taxonomic information.}

\item{grouped_by}{Character. The column name containing groups to summarize by (default: \code{"Species"}).}

\item{region_col}{Character. Column name for the region in which specimens were collected (default: \code{"Region"}).}

\item{institution_col}{Character. Column name for the institution or collection abbreviations (default: \code{"Institution"}).}

\item{number_col}{Character. Column name for specimen numbers (default: \code{"Number"}).}

\item{collector_col}{Character. Column name for the collector's name (default: \code{"Collector"}).}

\item{locality_col}{Character. Column name for locality names (default: \code{"Locality"}).}

\item{date_collected_col}{Character. Column name for collection dates (default: \code{"DateCollected"}).}

\item{type_col}{Character. Column name for specimen type designations (e.g., \code{"Holotype"}) (default: \code{"Type"}).}

\item{sex_col}{Character. Column name for sex information (default: \code{"Sex"}).}

\item{lat_col}{Character. Column name for latitude coordinates (default: \code{"Lat"}).}

\item{long_col}{Character. Column name for longitude coordinates (default: \code{"Long"}).}

\item{elevation_col}{Character. Column name for elevation information (default: \code{"Elevation"}).}
}
\value{
Prints a "Material Examined" section for each group to the console, formatted with specimen details grouped by region.
}
\description{
This function processes taxonomic and specimen data to create a detailed, well-formatted "Material Examined" section for to match descriptions. It handles missing columns by automatically filling them with \code{"N/A"},
ensuring smooth execution and consistent formatting, however, the function works best when the dataset includes all the specified columns. Specimens are grouped by the user specified group and region, and the
function outputs a comprehensive summary of specimens' metadata.
}
\details{
The function checks for the presence of each column and automatically fills in missing columns with \code{"N/A"}. It groups specimens by group and region,
and generates a "Material Examined" section that lists details including collector, locality, geographic coordinates, and elevation. The output includes
a summary with the number of specimens for each group, followed by a detailed list of specimen information.
}
\examples{
data(data_morphodata)
document_material_examined(
  data = data_morphodata,
  grouped_by = "Species",
  lat_col = "Latitude",
  long_col = "Longitude")

}
