% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_trait_range_table.R
\name{document_trait_range_table}
\alias{document_trait_range_table}
\title{Summarize Trait Ranges by Grouping Variable}
\usage{
document_trait_range_table(
  data,
  grouped_by,
  trait_columns,
  include_counts = FALSE,
  include_averages = FALSE
)
}
\arguments{
\item{data}{A data frame containing the data to be analyzed.}

\item{grouped_by}{A character string specifying the name of the column containing the grouping variable (default: \code{"Species"}).}

\item{trait_columns}{A character vector of trait column names for which the range or single value will be summarized.}

\item{include_counts}{Logical. If \code{TRUE}, includes the count of non-missing observations for each trait (default: \code{FALSE}).}

\item{include_averages}{Logical. If \code{TRUE}, includes the average value for each trait (default: \code{FALSE}).}
}
\value{
A data frame with one row per group, containing the range (or single value) for each trait, and optionally the average value and/or the count of non-missing observations.
Prior to publishing, all en dashes in ranges should be replaced with em dashes.
}
\description{
This function calculates the minimum and maximum values (or single value) for each specified trait,
grouped by a specified column (e.g., "Species", "Clade"). It can optionally include the average value
and the count of observations for each trait.
}
\details{
The function calculates the minimum and maximum values for each trait within each grouping variable. If the minimum and maximum are the same, it returns a single value. If all values for a trait are missing, it returns \code{"N/A"}.

If \code{include_averages = TRUE}, the average value for each trait is appended in parentheses. If \code{include_counts = TRUE}, the count of non-missing observations is appended in parentheses.
}
\examples{
data(data_morphodata)
document_trait_range_table(
  data = data_morphodata,
  grouped_by = "Species",
  trait_columns = c("trait1", "trait2", "trait3"))
document_trait_range_table(
  data = data_morphodata,
  grouped_by = "Clade",
  trait_columns = c("trait1", "trait2", "trait3"),
  include_counts = TRUE,
  include_averages = TRUE)

}
