% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_convex_hull.R
\name{geo_convex_hull}
\alias{geo_convex_hull}
\title{Calculate Minimum Convex Hull Area for Each Group}
\usage{
geo_convex_hull(data, grouped_by, lat_column, long_column, area_unit = "km^2")
}
\arguments{
\item{data}{A data frame containing occurrence data with latitude and longitude coordinates, and a column for group identifiers.}

\item{grouped_by}{A string specifying the column name for grouping identifiers (e.g., "Species", "Clade").}

\item{lat_column}{A string specifying the column name for latitude values.}

\item{long_column}{A string specifying the column name for longitude values.}

\item{area_unit}{A string specifying the unit for area calculation. Defaults to "km^2", but other options include "m^2".}
}
\value{
A data frame with three columns:
\itemize{
\item \code{grouped_by}: The name of the group (e.g., "Species", "Clade").
\item \code{Convex_Hull_Area}: The area of the minimum convex hull for each group in the specified unit.
\item \code{geometry}: The geometry of the minimum convex hull for each group.
}
}
\description{
This function calculates the area of the minimum convex hull for each group (e.g., "Species" or "Clades") based on
their geographic coordinates (latitude and longitude). The areas are returned in a data frame with group names
and their convex hull areas.
}
\details{
Groups with fewer than three occurrences are excluded from the calculation. The \code{sf} package is used to calculate
convex hulls and their areas. The \code{geometry} column contains the spatial geometry of the convex hull, which can be
used for further spatial analysis or visualization.
}
\examples{
data(data_morphodata)
geo_convex_hull(
  data = data_morphodata,
  area_unit = "km^2",
  grouped_by = "Clade",
  lat_column = "Latitude",
  long_column = "Longitude")

}
