% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_plot_map.R
\name{geo_plot_map}
\alias{geo_plot_map}
\title{Plot Occurrences and Convex Hulls on a World Map}
\usage{
geo_plot_map(data, grouped_by, lat_column, long_column)
}
\arguments{
\item{data}{A data frame containing occurrence data with columns for latitude and longitude, and a column for group identifiers (e.g., "Species", "Clade").}

\item{grouped_by}{A string specifying the column name for grouping identifiers.}

\item{lat_column}{A string specifying the column name for latitude values.}

\item{long_column}{A string specifying the column name for longitude values.}
}
\value{
An interactive \code{plotly} object displaying occurrences and their convex hulls on a world map.
}
\description{
This function visualizes occurrences and their convex hulls on a world map using \code{ggplot2}.
Geographic coordinates (latitude and longitude) are used to generate a map with occurrences as points
and convex hulls for groups with at least three occurrences. The map is interactive using the \code{plotly} package.
}
\details{
This function loads a world basemap from the \code{rnaturalearth} package and filters the data to remove rows with missing values.
It uses the \code{sf} package to create spatial objects and calculates convex hulls for groups with at least three occurrences.
The map is interactive via \code{plotly}, enabling zooming and hover-over features.
}
\examples{
data(data_morphodata)
geo_plot_map(
  data = data_morphodata,
  grouped_by = "Species",
  lat_column = "Latitude",
  long_column = "Longitude")

}
