% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_range.R
\name{geo_range}
\alias{geo_range}
\title{Calculate Geographic Range for Groups}
\usage{
geo_range(data, grouped_by, lat_column, long_column)
}
\arguments{
\item{data}{A data frame containing species and geographic data.}

\item{grouped_by}{A string specifying the column name for grouping identifiers (e.g., "Species" or "Clade").}

\item{lat_column}{A string specifying the name of the column containing latitude values.}

\item{long_column}{A string specifying the name of the column containing longitude values.}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{grouped_by}: Unique species or clades.
\item \code{Count}: Number of individuals for the group.
\item \code{Min_Lat}: Minimum latitude for the group.
\item \code{Max_Lat}: Maximum latitude for the group.
\item \code{Min_Long}: Minimum longitude for the group.
\item \code{Max_Long}: Maximum longitude for the group.
}
}
\description{
This function calculates the minimum and maximum latitude and longitude values for each species (or group) in the dataset,
and also counts the number of individuals per group.
}
\examples{
data(data_morphodata)
geo_range(
  data = data_morphodata,
  grouped_by = "Species",
  lat_column = "Latitude",
  long_column = "Longitude")
}
