% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_range_overlaps.R
\name{geo_range_overlaps}
\alias{geo_range_overlaps}
\title{Calculate Convex Hull Overlaps Between Groups}
\usage{
geo_range_overlaps(data, group_by, lat_col, long_col)
}
\arguments{
\item{data}{A data frame containing geographic data, including latitude and longitude coordinates.}

\item{group_by}{A string specifying the column name for grouping identifiers (e.g., "Species", "Clade").}

\item{lat_col}{A string specifying the column name for latitude values.}

\item{long_col}{A string specifying the column name for longitude values.}
}
\value{
A matrix indicating whether the convex hulls of groups overlap.
}
\description{
This function calculates convex range overlaps between groups (e.g., "Species" or "Clades") based on geographic coordinates.
The function generates convex hulls for groups with at least three coordinates by generating convex hulls, while also handling groups
with fewer than three points.
}
\details{
Groups with at least three coordinates generate convex hulls. For groups with fewer than three points,
special handling includes checking single points for containment and treating two points as a line to check for overlaps.
}
\examples{
data(data_morphodata)
geo_range_overlaps(
  data = data_morphodata,
  group_by = "Species",
  lat_col = "Latitude",
  long_col = "Longitude")

}
