% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_relatedness.R
\name{geo_relatedness}
\alias{geo_relatedness}
\title{Map Geographic Distribution with Phylogenetic Relatedness}
\usage{
geo_relatedness(
  data,
  lat_column,
  long_column,
  phylogeny,
  reference_species,
  individuals_column = "Species_updated"
)
}
\arguments{
\item{data}{A data frame containing geographic information for species. The data must include the columns specified in \code{lat_column} and \code{long_column} for latitude and longitude.}

\item{lat_column}{A string specifying the name of the column in \code{data} that contains latitude values.}

\item{long_column}{A string specifying the name of the column in \code{data} that contains longitude values.}

\item{phylogeny}{A phylogenetic tree of class \code{phylo} representing the evolutionary relationships between the species in the dataset.}

\item{reference_species}{A string specifying the name of the reference species in the phylogeny. This species will be used to color other species based on their relatedness to it.}

\item{individuals_column}{A string specifying the name of the column in \code{data} that contains individual species information. Defaults to "Species_updated".}
}
\value{
An interactive \code{plotly} plot showing the geographic distribution of species with phylogenetic relatedness colored by distance to the reference species.
}
\description{
This function generates an interactive map that visualizes the geographic distribution of species along with their phylogenetic relatedness to a reference species. Species are colored according to their relatedness to the specified reference species, with darker colors representing a higher degree of relatedness. The map is interactive, allowing for zooming and panning.
}
\details{
The function calculates pairwise phylogenetic distances using the \code{cophenetic} function, assigns relatedness values to species based on their distance to the reference species, and uses the \code{viridis} color palette to color the species according to their relatedness. The map includes a world basemap from Natural Earth and allows zooming and panning.
}
\examples{
# Example usage:
geo_relatedness(
  data = data_morphodata,
  lat_column = "Latitude",
  long_column = "Longitude",
  phylogeny = data_treeML,
  reference_species = "clade_1_outgroup",
  individuals_column = "Species_updated")

}
