% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_divergence_histogram.R
\name{phylo_divergence_histogram}
\alias{phylo_divergence_histogram}
\title{Plot Divergence Time Histogram for a Time Tree}
\usage{
phylo_divergence_histogram(time_tree, bin_width = 5, color = "lightblue")
}
\arguments{
\item{time_tree}{An object of class "phylo" or "timeTree", which contains the edge lengths representing divergence times.
If the tree does not contain edge lengths, an error will be raised.}

\item{bin_width}{Numeric value specifying the width of bins for the histogram. Default is 5.}

\item{color}{Character string specifying the color of the histogram bars. Default is "lightblue".}
}
\value{
The function creates a histogram plot that shows the distribution of divergence times
across the tree's nodes.
}
\description{
This function visualizes the distribution of divergence times for a given timetree.
It calculates the divergence times based on the edge lengths of the tree and then plots a histogram
to display the distribution of divergence times across the tree's nodes.
}
\details{
The function first checks whether the provided tree contains the \code{edge.length} attribute. If the attribute is missing,
the function stops and returns an error. If the values are present, a histogram is generated to visualize the distribution
of these divergence times, with the x-axis representing divergence time (in millions of years) and the y-axis representing
the frequency of nodes within specific time ranges.
}
\examples{
data(data_treeTime)
phylo_divergence_histogram(
  time_tree = data_treeTime,
  bin_width = 2,
  color = "darkseagreen")

}
