% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_heatmap.R
\name{phylo_heatmap}
\alias{phylo_heatmap}
\title{Create a Heatmap of Phylogenetic Relatedness}
\usage{
phylo_heatmap(phylogeny)
}
\arguments{
\item{phylogeny}{A phylogenetic tree object of class \code{phylo}. The tree must include tip labels representing the species.}
}
\value{
A \code{ggplot} object containing the heatmap of pairwise phylogenetic relatedness.
}
\description{
This function generates a heatmap visualizing the pairwise phylogenetic relatedness between species based on their phylogenetic tree. The relatedness is calculated using the cophenetic distances, which represent the phylogenetic distance between pairs of species.
}
\details{
The function first calculates pairwise phylogenetic distances using the \code{cophenetic()} function. The distances are then visualized as a heatmap using \code{ggplot2}. The axes of the heatmap are labeled with species names, and the color scale represents the phylogenetic distance between species.
}
\examples{
# Example usage:
phylo_heatmap(phylogeny = data_treeML)

}
