% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_highlight_taxa.R
\name{phylo_highlight_taxa}
\alias{phylo_highlight_taxa}
\title{Highlight Specific Taxa or Clades in Phylogenetic Trees}
\usage{
phylo_highlight_taxa(
  trees,
  taxa_to_highlight = NULL,
  highlight_colors = NULL,
  node_support_threshold = NULL,
  support_color = "blue",
  highlight_shape = "circle",
  shape_size = 1,
  tip_label_size = 0.8
)
}
\arguments{
\item{trees}{A list of phylogenetic trees of class \code{phylo}.}

\item{taxa_to_highlight}{A vector of taxa names (or patterns, such as regular expressions) to highlight in the trees. Default is \code{NULL}.}

\item{highlight_colors}{A vector of colors to use for highlighting the specified taxa. Default is \code{NULL}, which uses red.}

\item{node_support_threshold}{Numeric value specifying the minimum node support value for highlighting nodes. Default is \code{NULL}.}

\item{support_color}{Color used for highlighting nodes with high support values. Default is \code{"blue"}.}

\item{highlight_shape}{Shape of the highlight marker. Options are \code{"circle"} (default) or \code{"square"}.}

\item{shape_size}{Numeric value controlling the size of the highlight markers. Default is \code{2}.}

\item{tip_label_size}{Numeric value controlling the size of tip labels. Default is \code{0.8}.}
}
\value{
The function creates plots of the phylogenetic trees with highlighted taxa and/or nodes.
}
\description{
This function visualizes one or more phylogenetic trees with options to highlight specific taxa, clades, or nodes
based on a support threshold. Custom colors and shapes can be applied to tips or nodes.
}
\examples{
data(data_treeML, data_treeBayesian)
phylo_highlight_taxa(
    trees = list(data_treeML, data_treeBayesian),
    taxa_to_highlight = c("species_e", "species_d", "_a_"),
    highlight_colors = c("green", "orange", "purple"),
    node_support_threshold = 95,
    support_color = "blue",
    highlight_shape = "square",
    shape_size = 1,
    tip_label_size = 1.0)
}
