% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_merge_nodes.R
\name{phylo_merge_nodes}
\alias{phylo_merge_nodes}
\title{Merge Node Information from Multiple Trees}
\usage{
phylo_merge_nodes(ml_node_table, bayesian_node_table)
}
\arguments{
\item{ml_node_table}{A data frame containing node information from a Maximum Likelihood (ML) phylogenetic tree.
The table must include a column labeled "Taxa" representing the taxa names, and columns for node-specific data, including the support values.
Thistable can be generated using phylo_node_table().}

\item{bayesian_node_table}{A data frame containing node information from a Bayesian phylogenetic tree, with the same structure as the ML table,
including a "Taxa" column for taxa names and support values. This table can be generated using phylo_node_table().}
}
\value{
A data frame containing merged node information with columns:
\itemize{
\item \code{Taxa}: The taxa names.
\item \code{Node_ML}: The node information from the Maximum Likelihood tree.
\item \code{Node_Bayesian}: The node information from the Bayesian tree.
\item \code{Support_Values}: A string combining the support values from both trees, formatted as "ML_support/Bayesian_support".
}

The function also prints the resulting table.
}
\description{
This function merges node information from two phylogenetic trees (one Maximum Likelihood and one Bayesian) by matching taxa.
It combines the support values from both trees and creates a cleaner table with node and taxa information, including support values for each node.
Node tables can be generated using the \code{phylo_node_table()} function.
}
\details{
The function ensures that the "Taxa" column is treated as a character in both input tables. It merges the two tables by the "Taxa" column,
aligning nodes from both trees and adding suffixes to differentiate the columns from each tree. The support values from each tree are formatted as a
combined string, displaying both the Maximum Likelihood and Bayesian support values for each node. A final cleaned table is returned, containing the "Taxa",
"Node_ML", "Node_Bayesian", and "Support_Values" columns.
}
\examples{
# Using node tables made with generate_node_table().
# Output can be used with other functions (plot_tree_with_node_info()).
data(data_ml_node_table, data_bayesian_node_table)
data_merged_table <- phylo_merge_nodes(
  ml_node_table = data_ml_node_table,
  bayesian_node_table = data_bayesian_node_table)

}
