% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_node_support_hist.R
\name{phylo_node_support_hist}
\alias{phylo_node_support_hist}
\title{Plot Histogram of Node Support Values for ML and/or Bayesian Trees}
\usage{
phylo_node_support_hist(
  tree_ml = NULL,
  tree_bayesian = NULL,
  bin_width = 5,
  colors = c("skyblue", "green")
)
}
\arguments{
\item{tree_ml}{An object of class "phylo" containing a Maximum Likelihood (ML) tree with node support values.
Default is NULL.}

\item{tree_bayesian}{An object of class "phylo" containing a Bayesian tree with node support values.
Default is NULL.}

\item{bin_width}{Numeric value specifying the width of bins for the histogram. Default is 5.}

\item{colors}{A vector of colors to use for the histogram bars. If two trees are used the length of this vector should be 2.}
}
\value{
A bar plot displaying the distribution of node support values for the given tree(s).
}
\description{
This function generates a histogram of node support values from Maximum Likelihood (ML) trees, Bayesian trees, or both.
It visually compares the distribution of support values between the two types of trees. The support values are divided into
bins for easy visualization. If both trees are provided, their bars are displayed side by side for comparison.
If only one tree type is provided, its bars span the full bin width.
}
\details{
The function extracts node support values from the provided ML and/or Bayesian trees and stores them in a data frame.
The support values are then grouped into bins and visualized in a histogram. If both ML and Bayesian trees are provided,
their support value distributions are plotted side by side. If only one type of tree is provided, its bars fill the entire
bin width to maximize visibility.

The histogram uses a bar plot, with support value ranges on the x-axis and the frequency of nodes within each range on the y-axis.
Colors differentiate ML and Bayesian support values if both types of trees are provided. A legend is included for clarity.
}
\examples{
data(data_treeML, data_treeBayesian)
phylo_node_support_hist(
  tree_ml = data_treeML,
  tree_bayesian = data_treeBayesian,
  bin_width = 5,
  colors = c("darkorchid", "salmon"))

}
