% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_node_table.R
\name{phylo_node_table}
\alias{phylo_node_table}
\title{Generate a Table of Node Values from a Phylogenetic Tree}
\usage{
phylo_node_table(tree, method)
}
\arguments{
\item{tree}{A phylogenetic tree of class \code{phylo}.}

\item{method}{A string specifying the method for the node table. \code{"ML"},
\code{"Bayesian"}, and \code{"Time Tree"} are all accepted methods.}
}
\value{
A data frame containing the following columns:
\itemize{
\item \code{Node}: The node ID.
\item \code{Taxa}: A comma-separated list of descendant taxa.
\item \code{Support_or_Time}: The node's support value or divergence time.
}
}
\description{
This function creates a table of node values from a given phylogenetic tree,
including node IDs, descendant taxa, and either support values or divergence times,
depending on the specified method.
}
\details{
This function iterates over all internal nodes in the phylogenetic tree,
calculates the descendant taxa for each node, and retrieves the corresponding
support values or divergence times based on the specified method. If the \code{"Time Tree"} method is chosen, the function computes divergence
times using node ages derived from the tree's edge lengths.
}
\examples{
# Example: Generate a node table for a ML tree
phylo_node_table(
  tree = data_treeML,
  method = "ML")

}
