% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_plot_tree_nodes.R
\name{phylo_plot_tree_nodes}
\alias{phylo_plot_tree_nodes}
\title{Plot Phylogenetic Tree with Annotated Node Support Values}
\usage{
phylo_plot_tree_nodes(tree, tree_type = "ML", merged_table)
}
\arguments{
\item{tree}{An object of class \code{phylo} representing the phylogenetic tree to be plotted. The tree should contain node information for annotation.}

\item{tree_type}{A character string indicating the type of tree being plotted. Must be either \code{"ML"} or \code{"Bayesian"}.}

\item{merged_table}{A data frame containing node annotations, such as support values. This can be generated using \code{merge_node_tables()}.}
}
\value{
The function generates a plot with annotated node support values.
}
\description{
This function visualizes a phylogenetic tree with node support values from both Maximum Likelihood (ML) and Bayesian analyses.
It annotates tree nodes with support values from a provided data table and includes a subtitle indicating that bootstraps are displayed as ML/Bayesian.
A merged node table can be generated using the function \code{phylo_merge_nodes()}.
}
\details{
The function takes a phylogenetic tree (\code{tree}) and a table of node annotations (\code{merged_table}) generated from multiple analyses and annotates the tree nodes with
the corresponding support values. The tree is plotted using the \code{plot.phylo} function from the \code{ape} package, and node support values
are displayed using \code{nodelabels()}.

A subtitle ("Bootstraps are displayed as ML/Bayesian") is added below the main title to clarify the annotation format.
}
\examples{
# Assuming merged_table is available
phylo_plot_tree_nodes(
  tree = data_treeML,
  tree_type = "ML",
  merged_table = data_merged_table)

}
