% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_plot_trees.R
\name{phylo_plot_trees}
\alias{phylo_plot_trees}
\title{Plot Phylogenies}
\usage{
phylo_plot_trees(
  trees,
  methods,
  show_bootstraps = FALSE,
  bootstrap_color = "colored",
  show_posterior = FALSE,
  posterior_color = "colored",
  time_scale = FALSE,
  side_by_side = FALSE,
  name_data = NULL,
  phylogeny_type = "phylogram",
  edge_color = "black"
)
}
\arguments{
\item{trees}{A list of phylogenetic trees of class \code{phylo} or \code{multiPhylo} to be plotted.}

\item{methods}{A character vector of methods corresponding to the trees (e.g., "ML", "Bayesian", "Time Tree").}

\item{show_bootstraps}{Logical, indicating whether to display bootstrap values for ML trees. Default is FALSE.}

\item{bootstrap_color}{Color scheme for displaying bootstrap values. Default is "colored". Options are "none", "black", or "colored".}

\item{show_posterior}{Logical, indicating whether to display posterior support values for Bayesian trees. Default is FALSE.}

\item{posterior_color}{Color scheme for displaying posterior values. Default is "colored". Options are "none", "black", or "colored".}

\item{time_scale}{Logical, indicating whether to display divergence times for "Time Tree" methods. Default is FALSE.}

\item{side_by_side}{Logical, indicating whether to display trees side by side in a single plot. Default is FALSE.}

\item{name_data}{A data frame with two columns: Old_Name and New_Name, used for renaming tree tips. Default is NULL.}

\item{phylogeny_type}{The type of tree to plot: "phylogram" (default), "cladogram", or "radial".}

\item{edge_color}{Color for tree edges. Default is "black".}
}
\value{
This function does not return any values. It produces a plot of the phylogenetic trees with customizable features.
}
\description{
This function plots multiple phylogenetic trees, allowing customization options for plot type, support value display, branch color, and other visual elements.
When support values are "colored," values that are 95-100 are green, 85-94 are orange, and those below 85 are red.
}
\details{
The function can handle multiple trees and display them side by side.
The trees can be plotted as phylograms, cladograms, or radial trees.
Node values (e.g., bootstrap, posterior probabilities, or divergence times) can be displayed based on the selected method.
}
\examples{
data(data_treeML, data_treeBayesian, data_treeTime, data_names)
phylo_plot_trees(
  trees = list(data_treeML, data_treeBayesian, data_treeTime),
  methods = c("ML", "Bayesian", "Time Tree"),
  show_bootstraps = TRUE,
  bootstrap_color = "colored",
  show_posterior = TRUE,
  posterior_color = "black",
  time_scale = TRUE,
  side_by_side = TRUE,
  name_data = data_names)

}
