% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_taxa_divergence_hist.R
\name{phylo_taxa_divergence_hist}
\alias{phylo_taxa_divergence_hist}
\title{Plot Divergence Times Histogram for Specified Taxa}
\usage{
phylo_taxa_divergence_hist(tree, taxa_list, bin_width = 5, colors = NULL)
}
\arguments{
\item{tree}{A phylogenetic tree of class \code{phylo} that includes edge lengths.}

\item{taxa_list}{A character vector specifying the taxa or clades to analyze. Partial string matching is used.}

\item{bin_width}{A numeric value specifying the width of the bins for the histogram (default is 5).}

\item{colors}{A character vector of colors to use for each taxon in the plot. If not provided, the function defaults to a rainbow color palette.}
}
\value{
A bar plot showing the frequency distribution of divergence times for the specified taxa.
}
\description{
This function generates a histogram comparing the divergence times of multiple specified taxa
based on a phylogenetic tree. The taxa are matched using partial string matching to
identify nodes in the tree. Users can specify custom colors for each taxon.
}
\details{
The function computes node ages using edge lengths from the input tree, identifies
nodes corresponding to the specified taxa using partial string matching, and extracts their
divergence times. The divergence times are then binned, and a bar plot is created to compare
the distributions for the specified taxa. Each taxon is represented with a distinct color.

If no valid divergence times are found for any of the specified taxa, an error is returned.
}
\examples{
data(data_treeTime)
phylo_taxa_divergence_hist(
  tree = data_treeTime,
  taxa_list = c("clade_2", "clade_3", "_a"),
  colors = c("darkseagreen", "deeppink", "pink"),
  bin_width = 2)

}
