% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_tree_summary.R
\name{phylo_tree_summary}
\alias{phylo_tree_summary}
\title{Summarize Tree Statistics}
\usage{
phylo_tree_summary(tree, method)
}
\arguments{
\item{tree}{A phylogenetic tree of class 'phylo' or 'multiPhylo'.}

\item{method}{A character string indicating the method used to generate the tree. Options are "ML" (Maximum Likelihood), "Bayesian", or "Time Tree".}
}
\value{
A list with the following elements:
\itemize{
\item \code{num_nodes}: The number of internal nodes in the tree.
\item \code{max_depth}: The maximum depth of the tree (longest path from root to tip).
\item \code{bootstrap_summary}: A list of summary statistics (min, max, mean, mode, median, sd) for bootstrap values, if applicable.
\item \code{posterior_summary}: A list of summary statistics (min, max, mean, mode, median, sd) for posterior probabilities, if applicable.
\item \code{divergence_summary}: A list of summary statistics (min, max, mean, mode, median, sd) for divergence times, if applicable.
}
}
\description{
This function calculates summary statistics for a phylogenetic tree, including the number of nodes, depth of the tree, and additional statistics based on the tree's method (Maximum Likelihood, Bayesian, or Time Tree). It can summarize bootstrap values, posterior probabilities, or divergence times.
}
\details{
This function computes various summary statistics for a phylogenetic tree based on the specified method. For ML trees, it summarizes bootstrap values; for Bayesian trees, it summarizes posterior probabilities; for time trees, it summarizes divergence times. The statistics returned for each of these include the minimum, maximum, mean, mode, median, and standard deviation.
}
\examples{
data(data_treeML, data_treeBayesian, data_treeTime)
phylo_tree_summary(
  tree = data_treeML,
  method = "ML")

}
