% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_check_missing.R
\name{process_check_missing}
\alias{process_check_missing}
\title{Check Missing Data}
\usage{
process_check_missing(data)
}
\arguments{
\item{data}{A data frame to analyze for missing values.}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{Column} The name of each column in the dataset.
\item \code{Missing_raw} The raw count of missing values in each column.
\item \code{Missing_proportion} The proportion of missing values relative to the total number of rows in the dataset.}
}
}
\description{
This function analyzes a dataset and provides a summary of missing data for each column.
It calculates both the raw count and the proportion of missing values for each trait.
}
\examples{
data(data_morphodata)
process_check_missing(data = data_morphodata)

}
