% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_column_names.R
\name{process_column_names}
\alias{process_column_names}
\title{Rename Columns in a Dataset}
\usage{
process_column_names(data, new_column_names)
}
\arguments{
\item{data}{A data frame whose columns need to be renamed.}

\item{new_column_names}{A character vector containing the new column names. The length of this vector
should match the number of columns in the dataset.}
}
\value{
The updated data frame with renamed columns.
}
\description{
This function allows you to rename all columns in a given dataset. It ensures that the number of
new column names matches the number of columns in the dataset before applying the changes.
}
\note{
The number of new column names must exactly match the number of columns in the data frame.
If the lengths do not match, the function will print an error message and not perform any changes.
}
\examples{
# Rename columns in the dataset
process_column_names(
  data_morphodata,
  new_column_names = c("Clade", "Species", "Species Updated",
   "Institution", "Number", "Country", "Region", "Locality",
   "Latitude", "Longitude", "Elevation", "Trait 1", "Trait 2",
       "Trait 3", "Trait 4", "Trait 5", "Trait 6", "Trait 7",
       "Trait 8", "Trait 9", "Trait 10"))

}
