% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_convert_units.R
\name{process_convert_units}
\alias{process_convert_units}
\title{Convert Trait Units and Add a New Column}
\usage{
process_convert_units(data, trait_column, conversion_factor)
}
\arguments{
\item{data}{A data frame containing the trait data.}

\item{trait_column}{A character string specifying the name of the column to convert.}

\item{conversion_factor}{A numeric value to multiply the trait column values by for conversion.}
}
\value{
A modified data frame containing the original data and the new column with converted values.
}
\description{
This function applies a specified conversion factor to the values of a chosen trait column
and then adds the converted values as a new column in the dataset.
}
\details{
The function generates a new column in the dataset with the converted trait values.
The name of the new column is derived from the original trait column name,
appended with \code{"_converted"}.
}
\examples{
# Convert trait1 from cm to mm (conversion factor: 10)
data(data_morphodata)
process_convert_units(
  data = data_morphodata,
  trait_column = "trait1",
  conversion_factor = 10)

}
