% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_juveniles.R
\name{process_juveniles}
\alias{process_juveniles}
\title{Filter Out Juvenile Specimens Based on Trait Size}
\usage{
process_juveniles(data, trait, group_by, cutoff = 0.5)
}
\arguments{
\item{data}{A data frame containing morphological measurements.}

\item{trait}{A character string specifying the column name of the trait used for filtering (e.g., body size).}

\item{group_by}{A character string specifying the column name to group by.}

\item{cutoff}{A numeric value (default = 0.50) representing the proportion of the maximum trait
value in each group that individuals must meet or exceed to be retained. For example, a \code{cutoff = 0.90}
removes individuals that are smaller than 90\% of the largest individual in their group.}
}
\value{
A list containing:
\item{filtered_data}{A data frame with only individuals that meet the cutoff criteria.}
\item{removed_summary}{A data frame summarizing the number of individuals removed per group.}
}
\description{
This function removes individuals from a dataset that do not fall within a given percentage
of the largest specimen's trait value in each group (e.g., "species", "clade", ect.). It helps
filter out potential juveniles based on a specified trait.
}
\examples{
# Remove individuals that are less than 85\% of the largest trait1 value per species
process_juveniles(data_morphodata, trait = "trait1", group_by = "Species", cutoff = 0.85)

}
