% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_phylo_count.R
\name{process_phylo_count}
\alias{process_phylo_count}
\title{Count Specimens in a Phylogeny}
\usage{
process_phylo_count(data, group_name = NULL)
}
\arguments{
\item{data}{A phylogeny in the form of a 'phylo' object, typically from the \code{ape} package.}

\item{group_name}{A character string representing the name to match. Default is NULL.}
}
\value{
A list containing:
\item{total_specimens}{The total number of tips in the phylogeny.}
\item{specified_count}{The count of tips that match the specified character string. This is NULL
if no group_name is provided.}
}
\description{
This function takes a phylogeny as input and returns the total number of specimens
(tips) in the phylogeny, along with the count of specimens belonging to a user-specified
group. The group can be specified by a whole or partial character string.
}
\examples{
# Example usage with a phylogenetic tree (data_treeML should be a phylo object):
process_phylo_count(
  data = data_treeML,
  group_name = "clade_3")

}
