% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_phylo_names.R
\name{process_phylo_names}
\alias{process_phylo_names}
\title{Update Phylogeny Tree Tip Labels}
\usage{
process_phylo_names(tree, name_table)
}
\arguments{
\item{tree}{A phylogenetic tree of class \code{phylo}. The tree must have tip labels that match those in the \code{Old_Name} column of \code{name_table}.}

\item{name_table}{A data frame with two columns: \code{Old_Name} containing the current names in the tree, and \code{New_Name} containing the updated names.}
}
\value{
A \code{phylo} object with updated tip labels.
}
\description{
This function updates the tip labels in a phylogenetic tree based on a mapping table of old and new names.
It matches the names in the tree to those in the table and replaces the old names with the new ones.
}
\details{
The function assumes that the \code{name_table} has two columns: \code{Old_Name} and \code{New_Name}. It checks if the \code{tree}'s tip labels exist in the \code{Old_Name} column of \code{name_table} and updates the labels accordingly. If a name does not exist in the \code{name_table}, it remains unchanged.
}
\examples{
# Example usage:
process_phylo_names(data_treeML, data_names)

}
