% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_specimen_count.R
\name{process_specimen_count}
\alias{process_specimen_count}
\title{Count the Number of Occurrences per Group}
\usage{
process_specimen_count(data, grouped_by)
}
\arguments{
\item{data}{A data frame containing the data to be analyzed.}

\item{grouped_by}{A character string specifying the name of the column containing the grouping variable.}
}
\value{
A data frame with two columns:
\itemize{
\item \code{Group}: The group names (e.g., "species", "clade").
\item \code{Specimen_Count}: The number of occurrences (specimens) for each group. An additional row labeled "Total" is included at the end, representing the sum of all specimen counts.}
}
}
\description{
This function counts the number of occurrences (specimens) for each group (e.g., "species", "clade") in a dataset.
It also computes a total count for all groups combined.
}
\details{
The function groups the data by the specified grouping variable and counts the number of occurrences for each group.
It then adds a "Total" row with the sum of all the counts. The output data frame contains the group names, their corresponding counts, and the total count at the end.
}
\examples{
data(data_morphodata)
process_specimen_count(
  data = data_morphodata,
  grouped_by = "Species")

}
