% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_subset_group.R
\name{process_subset_group}
\alias{process_subset_group}
\title{Subset Dataset by Groups and Save to Environment}
\usage{
process_subset_group(data, grouped_by, prefix = "subset_")
}
\arguments{
\item{data}{A data frame to be subset.}

\item{grouped_by}{A character string specifying the column name used for grouping.}

\item{prefix}{A character string to prefix the names of the subset variables. Default is \code{"subset_"}.}
}
\value{
No direct return value. Subsets are saved as variables in the R environment.
}
\description{
This function subsets a dataset by unique values in a specified grouping column (e.g., "Species", "Clade")
and assigns each subset as a new variable in the R environment. The names of
the new variables are prefixed with a user-specified string.
}
\examples{
data(data_morphodata)
process_subset_group(
  data = data_morphodata,
  grouped_by = "Species",
  prefix = "Subset_")

}
