% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_compare.R
\name{trait_compare}
\alias{trait_compare}
\title{Compare Traits Across Groups}
\usage{
trait_compare(data, group_by_column, group_values, traits)
}
\arguments{
\item{data}{A data frame containing group and trait information.}

\item{group_by_column}{A string specifying the column name to group by (e.g., "Species" or "Clade").}

\item{group_values}{A character vector specifying the groups to include in the comparison.}

\item{traits}{A character vector specifying the trait columns to compare.}
}
\value{
Two sets of graphics are produced:
\itemize{
\item \code{pair_plot}: A base R scatterplot matrix comparing all specified traits across the selected groups. This will appear in the "plots" window.
\item \code{box_plots}: A list of interactive boxplots, one for each trait. These will appear in the "Viewer" window.
}
}
\description{
This function compares specified traits across selected groups (e.g., "species", "clades") using a variety of visualizations.
It employs both base R scatterplot matrices and interactive boxplots using \code{ggplot2} and \code{plotly}.
}
\details{
The function filters the input dataset for the selected groups and traits, removes missing values,
and generates the following visualizations:
\itemize{
\item A scatterplot matrix comparing all specified traits across the selected groups. This will appear in the "plots" window.
\item Boxplots showing the distribution of each trait for the selected groups. These will appear in the "Viewer" window.
}

The scatterplot matrix is displayed using base R graphics, while the boxplots are created using \code{ggplot2} and converted to interactive \code{plotly} objects.
}
\examples{
data(data_morphodata)
trait_compare(
  data = data_morphodata,
  group_by_column = "Species",
  group_values = c("Species A", "Species B"),
  traits = c("trait1", "trait2"))

}
