% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_correlation.R
\name{trait_correlation}
\alias{trait_correlation}
\title{Calculate and Visualize Correlations Between Traits}
\usage{
trait_correlation(data, trait_columns)
}
\arguments{
\item{data}{A data frame containing the trait data.}

\item{trait_columns}{A character vector specifying the names of the trait columns in the data frame for which the correlation should be calculated.}
}
\value{
A correlation matrix (data frame) showing pairwise correlations between the specified traits. The correlation matrix is also used to generate the heatmap plot.
}
\description{
This function calculates the correlation matrix between specified traits in the data and displays a heatmap of the correlations.
}
\details{
The function calculates pairwise correlations between the specified trait columns using Pearson's correlation coefficient and handles missing data with the "pairwise.complete.obs" method. It then reshapes the correlation matrix into a format suitable for visualization with \code{ggplot2} and creates a heatmap of the correlation values. The heatmap colors range from blue (negative correlation) to red (positive correlation), with white representing no correlation.
}
\examples{
# Calculate and visualize correlations between traits
data(data_morphodata)
trait_correlation(
  data = data_morphodata,
  trait_columns = c("trait1", "trait2", "trait3", "trait4", "trait5"))

}
