% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_filter.R
\name{trait_filter}
\alias{trait_filter}
\title{Filter Individuals by Trait Threshold}
\usage{
trait_filter(data, individuals_column, trait_column, threshold, above = TRUE)
}
\arguments{
\item{data}{A data frame containing the individual and trait data.}

\item{individuals_column}{A string specifying the name of the column containing unique individual identifiers.}

\item{trait_column}{A string specifying the name of the column containing the trait values.}

\item{threshold}{A numeric value specifying the threshold for filtering.}

\item{above}{A logical value. If TRUE, filters individuals with trait values above the threshold.
If FALSE, filters individuals with trait values below the threshold. Default is TRUE.}
}
\value{
A vector of individual identifiers that meet the specified trait threshold.
}
\description{
This function filters a dataset to identify individuals that meet a specified threshold
for a given trait. The filtering can be done for values above or below the threshold.
}
\examples{
# Filter individuals with trait values below 8
data(data_morphodata)
trait_filter(
  data = data_morphodata,
  individuals_column = "Species_updated",
  trait_column = "trait1",
  threshold = 8,
  above = FALSE)

}
