% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_not_separated.R
\name{trait_not_separated}
\alias{trait_not_separated}
\title{Identify Groups That Cannot Be Distinguished by Any Trait}
\usage{
trait_not_separated(data, trait_columns, grouped_by)
}
\arguments{
\item{data}{A data frame containing group names and their respective trait values.}

\item{trait_columns}{A character vector specifying the columns of traits to be compared between groups.}

\item{grouped_by}{A character vector specifying the column with group designations (e.g., "Species", "Clade").}
}
\value{
The function does not return any value but prints a list of groups that could not be separated by any trait.
}
\description{
This function identifies groups (e.g., "Species", "Clade") that cannot be distinguished from each other based on the given morphological traits.
It compares the minimum and maximum values for each trait across groups and checks if there is any overlap in the ranges.
If overlap is found for any trait, the pair is considered indistinguishable by that trait.
}
\details{
The function works by calculating the minimum and maximum values for each trait per group, then comparing these ranges across groups.
If the ranges of a particular trait overlap between two groups, the pair is considered non-separated by that trait. The function prints the
groups that could not be separated by any trait in pairs.
}
\examples{
data(data_morphodata)
trait_not_separated(
  data = data_morphodata,
  trait_columns = c("trait1", "trait2"),
  grouped_by = "Species")

}
