% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_summary.R
\name{trait_summary}
\alias{trait_summary}
\title{Calculate Summary Statistics for a Trait}
\usage{
trait_summary(data, grouped_by, trait_column)
}
\arguments{
\item{data}{A data frame containing the data to be analyzed.}

\item{grouped_by}{A character string specifying the name of the column containing the grouping variable.}

\item{trait_column}{A character string specifying the name of the column containing the trait data to summarize.}
}
\value{
A data frame with one row per group, containing the mean, median, and standard deviation for the specified trait.
}
\description{
This function calculates the mean, median, and standard deviation for a specified trait, grouped by a grouping variable (e.g., "species", "clade").
}
\details{
The function groups the data by the specified grouping variable and computes summary statistics (mean, median, and standard deviation) for the specified trait. Missing values are ignored in the calculations.
}
\examples{
data(data_morphodata)
trait_summary(
  data = data_morphodata,
  grouped_by = "Species",
  trait_column = "trait1")

}
